changes in the structure of the program
