+++ TeamCode/src/main/java/pedroPathing/PPConstants/FConstants.java
@@ -1,9 +1,7 @@
-package pedroPathing.constants;
+package pedroPathing.PPConstants;
 
 import com.pedropathing.localization.Localizers;
 import com.pedropathing.follower.FollowerConstants;
-import com.pedropathing.util.CustomFilteredPIDFCoefficients;
-import com.pedropathing.util.CustomPIDFCoefficients;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
 public class FConstants {
