+++ TeamCode/src/main/java/pedroPathing/examples/ExampleFieldCentricTeleop.java
@@ -3,11 +3,12 @@ package pedroPathing.examples;
 import com.pedropathing.follower.Follower;
 import com.pedropathing.localization.Pose;
 import com.pedropathing.util.Constants;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import  com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import pedroPathing.constants.FConstants;
-import pedroPathing.constants.LConstants;
+import pedroPathing.PPConstants.FConstants;
+import pedroPathing.PPConstants.LConstants;
 
 /**
  * This is an example teleop that showcases movement and field-centric driving.
@@ -17,6 +18,7 @@ import pedroPathing.constants.LConstants;
  */
 
 @TeleOp(name = "Example Field-Centric Teleop", group = "Examples")
+@Disabled
 public class ExampleFieldCentricTeleop extends OpMode {
     private Follower follower;
     private final Pose startPose = new Pose(0,0,0);
