+++ TeamCode/src/main/java/pedroPathing/examples/ExampleRobotCentricTeleop.java
@@ -1,15 +1,14 @@
 package pedroPathing.examples;
 
-import android.provider.SyncStateContract;
-
 import com.pedropathing.follower.Follower;
 import com.pedropathing.localization.Pose;
 import com.pedropathing.util.Constants;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import  com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import pedroPathing.constants.FConstants;
-import pedroPathing.constants.LConstants;
+import pedroPathing.PPConstants.FConstants;
+import pedroPathing.PPConstants.LConstants;
 
 /**
  * This is an example teleop that showcases movement and robot-centric driving.
@@ -19,6 +18,7 @@ import pedroPathing.constants.LConstants;
  */
 
 @TeleOp(name = "Example Robot-Centric Teleop", group = "Examples")
+@Disabled
 public class ExampleRobotCentricTeleop extends OpMode {
     private Follower follower;
     private final Pose startPose = new Pose(0,0,0);
