--- TeamCode/src/main/java/pedroPathing/examples/trajectory-v2.pp
@@ -1,1 +1,0 @@
-{"startPoint":{"x":10,"y":66.5,"heading":"constant","degrees":0},"lines":[{"endPoint":{"x":38,"y":66.5,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[],"color":"#B978DD"},{"endPoint":{"x":60,"y":30,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[{"x":9.6,"y":20.7},{"x":62.7,"y":42.3}],"color":"#D955D6"},{"endPoint":{"x":20,"y":30,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[],"color":"#789A6A"},{"endPoint":{"x":60,"y":20,"heading":"linear","reverse":false,"endDeg":0,"startDeg":0},"controlPoints":[{"x":60,"y":30}],"color":"#C8BCAD"},{"endPoint":{"x":20,"y":20,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[],"color":"#B67DDC"},{"endPoint":{"x":60,"y":10,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[{"x":60,"y":20}],"color":"#A98A5A"},{"endPoint":{"x":20,"y":10,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[],"color":"#CBA88C"}]}
