--- TeamCode/src/main/java/pedroPathing/examples/trajectory.pp
@@ -1,1 +1,0 @@
-{"startPoint":{"x":10,"y":66.5,"heading":"linear","startDeg":0,"endDeg":0},"lines":[{"endPoint":{"x":38,"y":66.5,"heading":"linear","startDeg":0,"endDeg":0},"controlPoints":[],"color":"#79799D"},{"endPoint":{"x":55,"y":34.5,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[{"x":14.5,"y":36.5}],"color":"#759CAD"},{"endPoint":{"x":21,"y":31.5,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[{"x":60,"y":25}],"color":"#B789A8"},{"endPoint":{"x":21,"y":16.5,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[{"x":100,"y":14.5}],"color":"#7585B7"},{"endPoint":{"x":22,"y":11.5,"heading":"linear","reverse":false,"startDeg":0,"endDeg":0},"controlPoints":[{"x":94.5,"y":24.9},{"x":71.2,"y":0.6}],"color":"#786D59"},{"endPoint":{"x":22,"y":26.5,"heading":"linear","reverse":false,"startDeg":0,"endDeg":180},"controlPoints":[{"x":80,"y":16.5}],"color":"#5AACA8"},{"endPoint":{"x":20,"y":26.5,"heading":"linear","reverse":false,"startDeg":180,"endDeg":180},"controlPoints":[],"color":"#99B959"},{"endPoint":{"x":39,"y":67.5,"heading":"linear","reverse":false,"startDeg":180,"endDeg":0},"controlPoints":[],"color":"#A66DAD"},{"endPoint":{"x":22,"y":26.5,"heading":"linear","reverse":false,"startDeg":0,"endDeg":180},"controlPoints":[],"color":"#D8D9AB"},{"endPoint":{"x":20,"y":26.5,"heading":"linear","reverse":false,"startDeg":180,"endDeg":180},"controlPoints":[],"color":"#5C8659"},{"endPoint":{"x":39,"y":68.5,"heading":"linear","reverse":false,"startDeg":180,"endDeg":0},"controlPoints":[],"color":"#5BD98D"}]}
