+++ TeamCode/src/main/java/pedroPathing/simulation/FakeDcMotor.java
@@ -1,213 +1,213 @@
-package pedroPathing.simulation;
-
-import com.qualcomm.robotcore.hardware.DcMotorController;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.PIDCoefficients;
-import com.qualcomm.robotcore.hardware.PIDFCoefficients;
-import com.qualcomm.robotcore.hardware.configuration.typecontainers.MotorConfigurationType;
-
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.CurrentUnit;
-
-public class FakeDcMotor implements DcMotorEx {
-    private int position = 0;
-    private double power = 0;
-
-    @Override
-    public MotorConfigurationType getMotorType() {
-        return null;
-    }
-
-    @Override
-    public void setMotorType(MotorConfigurationType motorType) {
-
-    }
-
-    @Override
-    public DcMotorController getController() {
-        return null;
-    }
-
-    @Override
-    public int getPortNumber() {
-        return 0;
-    }
-
-    @Override
-    public void setZeroPowerBehavior(ZeroPowerBehavior zeroPowerBehavior) {
-
-    }
-
-    @Override
-    public ZeroPowerBehavior getZeroPowerBehavior() {
-        return null;
-    }
-
-    @Override
-    public void setPowerFloat() {
-
-    }
-
-    @Override
-    public boolean getPowerFloat() {
-        return false;
-    }
-
-    @Override public void setTargetPosition(int pos) { this.position = pos; }
-    @Override public int getTargetPosition() { return position; }
-
-    @Override
-    public boolean isBusy() {
-        return false;
-    }
-
-    @Override
-    public void setDirection(Direction direction) {
-
-    }
-
-    @Override
-    public Direction getDirection() {
-        return null;
-    }
-
-    @Override public void setPower(double power) { this.power = power; }
-    @Override public double getPower() { return power; }
-    @Override public int getCurrentPosition() { return position; }
-
-    @Override
-    public void setMode(RunMode mode) {
-
-    }
-
-    @Override
-    public RunMode getMode() {
-        return null;
-    }
-
-    @Override
-    public void setMotorEnable() {
-
-    }
-
-    @Override
-    public void setMotorDisable() {
-
-    }
-
-    @Override
-    public boolean isMotorEnabled() {
-        return false;
-    }
-
-    @Override
-    public void setVelocity(double angularRate) {
-
-    }
-
-    @Override
-    public void setVelocity(double angularRate, AngleUnit unit) {
-
-    }
-
-    @Override
-    public double getVelocity() {
-        return 0;
-    }
-
-    @Override
-    public double getVelocity(AngleUnit unit) {
-        return 0;
-    }
-
-    @Override
-    public void setPIDCoefficients(RunMode mode, PIDCoefficients pidCoefficients) {
-
-    }
-
-    @Override
-    public void setPIDFCoefficients(RunMode mode, PIDFCoefficients pidfCoefficients) throws UnsupportedOperationException {
-
-    }
-
-    @Override
-    public void setVelocityPIDFCoefficients(double p, double i, double d, double f) {
-
-    }
-
-    @Override
-    public void setPositionPIDFCoefficients(double p) {
-
-    }
-
-    @Override
-    public PIDCoefficients getPIDCoefficients(RunMode mode) {
-        return null;
-    }
-
-    @Override
-    public PIDFCoefficients getPIDFCoefficients(RunMode mode) {
-        return null;
-    }
-
-    @Override
-    public void setTargetPositionTolerance(int tolerance) {
-
-    }
-
-    @Override
-    public int getTargetPositionTolerance() {
-        return 0;
-    }
-
-    @Override
-    public double getCurrent(CurrentUnit unit) {
-        return 0;
-    }
-
-    @Override
-    public double getCurrentAlert(CurrentUnit unit) {
-        return 0;
-    }
-
-    @Override
-    public void setCurrentAlert(double current, CurrentUnit unit) {
-
-    }
-
-    @Override
-    public boolean isOverCurrent() {
-        return false;
-    }
-
-    @Override
-    public Manufacturer getManufacturer() {
-        return null;
-    }
-
-    @Override
-    public String getDeviceName() {
-        return "";
-    }
-
-    @Override
-    public String getConnectionInfo() {
-        return "";
-    }
-
-    @Override
-    public int getVersion() {
-        return 0;
-    }
-
-    @Override
-    public void resetDeviceConfigurationForOpMode() {
-
-    }
-
-    @Override
-    public void close() {
-
-    }
-}
+//package pedroPathing.simulation;
+//
+//import com.qualcomm.robotcore.hardware.DcMotorController;
+//import com.qualcomm.robotcore.hardware.DcMotorEx;
+//import com.qualcomm.robotcore.hardware.DcMotor;
+//import com.qualcomm.robotcore.hardware.PIDCoefficients;
+//import com.qualcomm.robotcore.hardware.PIDFCoefficients;
+//import com.qualcomm.robotcore.hardware.configuration.typecontainers.MotorConfigurationType;
+//
+//import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
+//import org.firstinspires.ftc.robotcore.external.navigation.CurrentUnit;
+//
+//public class FakeDcMotor implements DcMotorEx {
+//    private int position = 0;
+//    private double power = 0;
+//
+//    @Override
+//    public MotorConfigurationType getMotorType() {
+//        return null;
+//    }
+//
+//    @Override
+//    public void setMotorType(MotorConfigurationType motorType) {
+//
+//    }
+//
+//    @Override
+//    public DcMotorController getController() {
+//        return null;
+//    }
+//
+//    @Override
+//    public int getPortNumber() {
+//        return 0;
+//    }
+//
+//    @Override
+//    public void setZeroPowerBehavior(ZeroPowerBehavior zeroPowerBehavior) {
+//
+//    }
+//
+//    @Override
+//    public ZeroPowerBehavior getZeroPowerBehavior() {
+//        return null;
+//    }
+//
+//    @Override
+//    public void setPowerFloat() {
+//
+//    }
+//
+//    @Override
+//    public boolean getPowerFloat() {
+//        return false;
+//    }
+//
+//    @Override public void setTargetPosition(int pos) { this.position = pos; }
+//    @Override public int getTargetPosition() { return position; }
+//
+//    @Override
+//    public boolean isBusy() {
+//        return false;
+//    }
+//
+//    @Override
+//    public void setDirection(Direction direction) {
+//
+//    }
+//
+//    @Override
+//    public Direction getDirection() {
+//        return null;
+//    }
+//
+//    @Override public void setPower(double power) { this.power = power; }
+//    @Override public double getPower() { return power; }
+//    @Override public int getCurrentPosition() { return position; }
+//
+//    @Override
+//    public void setMode(RunMode mode) {
+//
+//    }
+//
+//    @Override
+//    public RunMode getMode() {
+//        return null;
+//    }
+//
+//    @Override
+//    public void setMotorEnable() {
+//
+//    }
+//
+//    @Override
+//    public void setMotorDisable() {
+//
+//    }
+//
+//    @Override
+//    public boolean isMotorEnabled() {
+//        return false;
+//    }
+//
+//    @Override
+//    public void setVelocity(double angularRate) {
+//
+//    }
+//
+//    @Override
+//    public void setVelocity(double angularRate, AngleUnit unit) {
+//
+//    }
+//
+//    @Override
+//    public double getVelocity() {
+//        return 0;
+//    }
+//
+//    @Override
+//    public double getVelocity(AngleUnit unit) {
+//        return 0;
+//    }
+//
+//    @Override
+//    public void setPIDCoefficients(RunMode mode, PIDCoefficients pidCoefficients) {
+//
+//    }
+//
+//    @Override
+//    public void setPIDFCoefficients(RunMode mode, PIDFCoefficients pidfCoefficients) throws UnsupportedOperationException {
+//
+//    }
+//
+//    @Override
+//    public void setVelocityPIDFCoefficients(double p, double i, double d, double f) {
+//
+//    }
+//
+//    @Override
+//    public void setPositionPIDFCoefficients(double p) {
+//
+//    }
+//
+//    @Override
+//    public PIDCoefficients getPIDCoefficients(RunMode mode) {
+//        return null;
+//    }
+//
+//    @Override
+//    public PIDFCoefficients getPIDFCoefficients(RunMode mode) {
+//        return null;
+//    }
+//
+//    @Override
+//    public void setTargetPositionTolerance(int tolerance) {
+//
+//    }
+//
+//    @Override
+//    public int getTargetPositionTolerance() {
+//        return 0;
+//    }
+//
+//    @Override
+//    public double getCurrent(CurrentUnit unit) {
+//        return 0;
+//    }
+//
+//    @Override
+//    public double getCurrentAlert(CurrentUnit unit) {
+//        return 0;
+//    }
+//
+//    @Override
+//    public void setCurrentAlert(double current, CurrentUnit unit) {
+//
+//    }
+//
+//    @Override
+//    public boolean isOverCurrent() {
+//        return false;
+//    }
+//
+//    @Override
+//    public Manufacturer getManufacturer() {
+//        return null;
+//    }
+//
+//    @Override
+//    public String getDeviceName() {
+//        return "";
+//    }
+//
+//    @Override
+//    public String getConnectionInfo() {
+//        return "";
+//    }
+//
+//    @Override
+//    public int getVersion() {
+//        return 0;
+//    }
+//
+//    @Override
+//    public void resetDeviceConfigurationForOpMode() {
+//
+//    }
+//
+//    @Override
+//    public void close() {
+//
+//    }
+//}
