+++ TeamCode/src/main/java/pedroPathing/simulation/TestableTeleOpMain.java
@@ -1,21 +1,21 @@
-package pedroPathing.simulation;
-
-import pedroPathing.examples.TeleOpMain;
-
-public class TestableTeleOpMain extends TeleOpMain {
-    private boolean active = false;
-
-    public void setActive(boolean active) {
-        this.active = active;
-    }
-
-//    @Override
-//    public boolean opModeIsActive() {
-//        return active;
+//package pedroPathing.simulation;
+//
+//import pedroPathing.examples.TeleOpMain;
+//
+//public class TestableTeleOpMain extends TeleOpMain {
+//    private boolean active = false;
+//
+//    public void setActive(boolean active) {
+//        this.active = active;
 //    }
-
-    @Override
-    public void waitForStart() {
-        // Simulate immediate start
-    }
-}
+//
+////    @Override
+////    public boolean opModeIsActive() {
+////        return active;
+////    }
+//
+//    @Override
+//    public void waitForStart() {
+//        // Simulate immediate start
+//    }
+//}
